- Moved the "More" functionality into the MovieImageController so that the examples are now in the same order as Tim presents them - Created a new "base" view controller for the MovieListViewController, and moved some "boiler plate" into this controller.  This should make MovieListViewController easier and cleaner for the demo
