+++ StrongReferenceDemo/Base.lproj/Main.storyboard
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="9532" systemVersion="15D21" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="HM7-1h-KCR">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="9532" systemVersion="15E65" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" initialViewController="HM7-1h-KCR">
     <dependencies>
         <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="9530"/>
@@ -60,13 +60,7 @@
                             <outlet property="delegate" destination="cN2-Ag-w7O" id="GVi-Si-XYp"/>
                         </connections>
                     </tableView>
-                    <navigationItem key="navigationItem" id="PNY-Xm-vwa">
-                        <barButtonItem key="rightBarButtonItem" title="More" id="Ndv-YS-rka">
-                            <connections>
-                                <action selector="handleMoreAction:" destination="cN2-Ag-w7O" id="2DB-NW-yAg"/>
-                            </connections>
-                        </barButtonItem>
-                    </navigationItem>
+                    <navigationItem key="navigationItem" id="PNY-Xm-vwa"/>
                     <connections>
                         <segue destination="OBj-tW-sDM" kind="show" identifier="navigateToMovie" id="qDX-5n-ZUo"/>
                     </connections>
