+++ StrongReferenceDemo/StringExtension.swift
@@ -1,0 +1,13 @@
+//
+//  StringUtil.swift
+//  StrongReferenceDemo
+//
+//  Created by Matthews, Jamie on 3/25/16.
+//  Copyright © 2016 Amazon. All rights reserved.
+//
+
+extension String {
+    func urlEncodedString() -> String {
+        return self.stringByAddingPercentEncodingWithAllowedCharacters(.URLHostAllowedCharacterSet())!
+    }
+}
